requirements.txt
# requirements.txt
git+https://github.com/illuin-tech/colpali
streamlit
torch
Pillow
numpy
pdf2image
setuptools
sqlite3